refactor(changelog-generator): extract readYarnrc() into separate file
Decluttering.
