+++ projects/npm-tools/packages/changelog-generator/src/index.js
@@ -9,6 +9,7 @@ const {promisify} = require('util');
 
 const {cleanup, error, info, log, prompt, warn} = require('./console');
 const git = require('./git');
+const readYarnrc = require('./readYarnrc');
 
 const readFileAsync = promisify(fs.readFile);
 const writeFileAsync = promisify(fs.writeFile);
@@ -637,7 +638,7 @@ async function getVersion(options) {
 			preid = options.preid;
 		}
 		else {
-			const settings = await getYarnRc();
+			const settings = await readYarnrc();
 
 			preid = settings.get('--version.preid') || preid || 'pre';
 		}
@@ -720,7 +721,7 @@ async function getVersion(options) {
  * wrong directory), an error is thrown.
  */
 async function getVersionTagPrefix() {
-	const settings = await getYarnRc();
+	const settings = await readYarnrc();
 
 	const setting = settings.get('version-tag-prefix');
 
@@ -733,76 +734,6 @@ async function getVersionTagPrefix() {
 	}
 
 	return '';
-}
-
-const COMMENT_REGEXP = /^\s*#/;
-
-/**
- * Reads any .yarnrc files between the repo root and in the current working
- * directory, and returns a map representing their contents.
- *
- * Comments are ignored.
- *
- * Because local settings should override ones higher up in the hierarchy,
- * duplicate settings are resolved by keeping only the last-seen instance of any
- * given setting.
- */
-async function getYarnRc() {
-	try {
-		fs.accessSync('package.json', fs.constants.R_OK);
-	}
-	catch (_error) {
-		throw new Error(
-			'Expected to run from a directory with a "package.json"'
-		);
-	}
-
-	const settings = new Map();
-
-	let candidate = process.cwd();
-
-	const root = (await git('rev-parse', '--show-toplevel')).trim();
-
-	// Will visit from most general to most local.
-
-	while (true) {
-		try {
-			const contents = await readFileAsync(
-				path.join(candidate, '.yarnrc'),
-				'utf8'
-			);
-
-			contents.split(/\r\n|\r|\n/).forEach((line) => {
-				if (COMMENT_REGEXP.test(line)) {
-					return;
-				}
-				else {
-					const match = line.match(/^\s*(\S+)\s+(.+)\s*$/);
-
-					if (match) {
-						settings.set(match[1], match[2]);
-					}
-				}
-			});
-		}
-		catch (_error) {
-
-			// No readable .yarnrc.
-
-		}
-
-		if (candidate === root) {
-			break;
-		}
-
-		const components = candidate.split(path.sep);
-
-		components.pop();
-
-		candidate = components.join(path.sep);
-	}
-
-	return settings;
 }
 
 async function main(_node, _script, ...args) {
